+++ app_build_suite/build_steps/pytest.py
@@ -2,7 +2,6 @@ import argparse
 import logging
 import os
 import shutil
-import subprocess  # nosec, needed to invoke pytest as external process
 from abc import ABC
 from typing import Set, cast, List
 
@@ -21,6 +20,7 @@ from app_build_suite.build_steps.helm import context_key_chart_file_name
 from app_build_suite.errors import ValidationError, TestError
 from app_build_suite.types import Context
 from app_build_suite.utils.config import get_config_value_by_cmd_line_option
+from app_build_suite.utils.processes import run_and_log
 
 logger = logging.getLogger(__name__)
 
@@ -103,7 +103,7 @@ class PytestTestRunner(BaseTestRunner, ABC):
             f"Running {self._pipenv_bin} tool in '{self._pytest_dir}' directory to install virtual env "
             f"for running tests."
         )
-        run_res = subprocess.run(args, cwd=self._pytest_dir)  # nosec, no user input here
+        run_res = run_and_log(args, print_debug=True, cwd=self._pytest_dir)  # nosec, no user input here
         if run_res.returncode != 0:
             raise TestError(f"Running '{args}' in directory '{self._pytest_dir}' failed.")
 
@@ -138,7 +138,7 @@ class PytestTestRunner(BaseTestRunner, ABC):
         if app_config_file_path:
             args += ["--values-file", app_config_file_path]
         logger.info(f"Running {self._pytest_bin} tool in '{self._pytest_dir}' directory.")
-        run_res = subprocess.run(args, cwd=self._pytest_dir)  # nosec, no user input here
+        run_res = run_and_log(args, print_debug=True, cwd=self._pytest_dir)  # nosec, no user input here
         if run_res.returncode != 0:
             raise TestError(f"Pytest tests failed: running '{args}' in directory '{self._pytest_dir}' failed.")